+++ db/migrate/20150904115735_add_content_id_to_attachments.rb
@@ -1,0 +1,5 @@
+class AddContentIdToAttachments < ActiveRecord::Migration
+  def change
+    add_column :attachments, :content_id, :string
+  end
+end