+++ db/schema.rb
@@ -11,7 +11,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20150828125947) do
+ActiveRecord::Schema.define(version: 20150904115735) do
 
   create_table "attachments", force: :cascade do |t|
     t.integer  "attachable_id"
@@ -22,6 +22,7 @@ ActiveRecord::Schema.define(version: 20150828125947) do
     t.string   "file_content_type"
     t.integer  "file_file_size"
     t.datetime "file_updated_at"
+    t.string   "content_id"
   end
 
   add_index "attachments", ["attachable_id"], name: "index_attachments_on_attachable_id"